+++ shaders/default_vert.glsl
@@ -1,5 +1,6 @@
+#version 120
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 //
 //
